Rename Jdk7DirectoryWalker to DefaultDirectoryWalker
