+++ subprojects/files/src/test/groovy/org/gradle/api/internal/file/collections/DirectoryWalkerTest.groovy
@@ -21,9 +21,7 @@ import org.gradle.api.JavaVersion
 import org.gradle.api.file.FileVisitDetails
 import org.gradle.api.file.FileVisitor
 import org.gradle.api.internal.file.TestFiles
-import org.gradle.api.internal.file.collections.jdk7.Jdk7DirectoryWalker
 import org.gradle.api.tasks.util.PatternSet
-import org.gradle.internal.Factory
 import org.gradle.test.fixtures.file.TestFile
 import org.gradle.util.Requires
 import org.gradle.util.UsesNativeServices
@@ -36,7 +34,10 @@ import java.util.concurrent.atomic.AtomicInteger
 class DirectoryWalkerTest extends AbstractDirectoryWalkerTest<DirectoryWalker> {
     @Override
     protected List<DirectoryWalker> getWalkers() {
-        return [new DefaultDirectoryWalker(), new Jdk7DirectoryWalker(), new ReproducibleDirectoryWalker()]
+        return [
+            new DefaultDirectoryWalker(),
+            new ReproducibleDirectoryWalker()
+        ]
     }
 
     // java.nio2 cannot access files with unicode characters when using single-byte non-unicode platform encoding
@@ -54,9 +55,9 @@ class DirectoryWalkerTest extends AbstractDirectoryWalkerTest<DirectoryWalker> {
         where:
         fileEncoding | expectedClassName
         "UTF-8"      | "Jdk7DirectoryWalker"
-        "UTF-16be" | "Jdk7DirectoryWalker"
-        "UTF-16le" | "Jdk7DirectoryWalker"
-        "UTF-16"   | "Jdk7DirectoryWalker"
+        "UTF-16be"   | "Jdk7DirectoryWalker"
+        "UTF-16le"   | "Jdk7DirectoryWalker"
+        "UTF-16"     | "Jdk7DirectoryWalker"
         "ISO-8859-1" | "DefaultDirectoryWalker"
     }
 
@@ -66,8 +67,8 @@ class DirectoryWalkerTest extends AbstractDirectoryWalkerTest<DirectoryWalker> {
         generateFilesAndSubDirectories(rootDir, 10, 5, 3, 1, new AtomicInteger(0))
 
         when:
-        def visitedWithJdk7Walker = walkFiles(rootDir, new Jdk7DirectoryWalker())
-        def visitedWithDefaultWalker = walkFiles(rootDir, new DefaultDirectoryWalker())
+        def visitedWithJdk7Walker = walkFiles(rootDir)
+        def visitedWithDefaultWalker = walkFiles(rootDir)
 
         then:
         visitedWithDefaultWalker.size() == 340
@@ -102,8 +103,8 @@ class DirectoryWalkerTest extends AbstractDirectoryWalkerTest<DirectoryWalker> {
         }
     }
 
-    private static List<FileVisitDetails> walkFiles(rootDir, walkerInstance) {
-        def fileTree = new DirectoryFileTree(rootDir, new PatternSet(), { walkerInstance } as Factory, TestFiles.fileSystem(), false)
+    private static List<FileVisitDetails> walkFiles(rootDir) {
+        def fileTree = new DirectoryFileTree(rootDir, new PatternSet(), TestFiles.fileSystem(), false)
         def visited = []
         def visitClosure = { visited << it }
         def fileVisitor = [visitFile: visitClosure, visitDir: visitClosure] as FileVisitor
@@ -114,15 +115,14 @@ class DirectoryWalkerTest extends AbstractDirectoryWalkerTest<DirectoryWalker> {
     def "file walker sees a snapshot of file metadata even if files are deleted after walking has started"() {
         given:
         def rootDir = tmpDir.createDir("root")
-        long minimumTimestamp = (System.currentTimeMillis()/1000 * 1000) - 2000
+        long minimumTimestamp = (System.currentTimeMillis() / 1000 * 1000) - 2000
         def file1 = rootDir.createFile("a/b/1.txt")
         file1 << '12345'
         def file2 = rootDir.createFile("a/b/2.txt")
         file2 << '12345'
         def file3 = rootDir.createFile("a/b/3.txt")
         file3 << '12345'
-        def walkerInstance = new Jdk7DirectoryWalker()
-        def fileTree = new DirectoryFileTree(rootDir, new PatternSet(), { walkerInstance } as Factory, TestFiles.fileSystem(), false)
+        def fileTree = new DirectoryFileTree(rootDir, new PatternSet(), TestFiles.fileSystem(), false)
         def visitedFiles = []
         def visitedDirectories = []
         def fileVisitor = [visitFile: { visitedFiles << it }, visitDir: { visitedDirectories << it }] as FileVisitor
@@ -147,7 +147,7 @@ class DirectoryWalkerTest extends AbstractDirectoryWalkerTest<DirectoryWalker> {
         def visited = []
         def visitClosure = { visited << it.file.absolutePath }
         def fileVisitor = [visitFile: visitClosure, visitDir: visitClosure] as FileVisitor
-        def fileTree = new DirectoryFileTree(rootDir, patternSet, { walkerInstance } as Factory, TestFiles.fileSystem(), false)
+        def fileTree = new DirectoryFileTree(rootDir, patternSet, TestFiles.fileSystem(), false)
         fileTree.visit(fileVisitor)
         return visited
     }
