+++ subprojects/files/src/testFixtures/groovy/org/gradle/api/internal/file/collections/AbstractDirectoryWalkerTest.groovy
@@ -18,7 +18,6 @@ package org.gradle.api.internal.file.collections
 
 import com.google.common.annotations.VisibleForTesting
 import org.gradle.api.file.FileVisitor
-import org.gradle.api.internal.file.collections.jdk7.Jdk7DirectoryWalker
 import org.gradle.api.tasks.util.PatternSet
 import org.gradle.internal.Factory
 import org.gradle.internal.nativeintegration.filesystem.FileSystem
@@ -215,7 +214,7 @@ abstract class AbstractDirectoryWalkerTest<T> extends Specification {
         file2 << '12345'
         def file3 = rootDir.createFile("a/b/3.txt")
         file3 << '12345'
-        def walkerInstance = new Jdk7DirectoryWalker()
+        def walkerInstance = new DefaultDirectoryWalker()
         def fileTree = new DirectoryFileTree(rootDir, new PatternSet(), { walkerInstance } as Factory, NativeServicesTestFixture.getInstance().get(FileSystem), false)
         def visitedFiles = []
         def visitedDirectories = []
